// src/data/types.ts
export interface ModificationRow {
  Id: number
  UnitId: number
  Name: string | null
  Type: number
  UIName: string | null
  ThumbnailFileName: string | null
  Order: number
}

export interface OptionRow {
  Id: number
  IsDefault: boolean
  ModificationId: number
  Name: string | null
  Order: number
  UIName: string | null
  ConcatenateWithUnitName: string | null
  ReplaceUnitName: string | null
  ReplaceUnitId: number
  Cost: number
  OptionPicture: string | null
  ThumbnailOverride: string | null
  PortraitOverride: string | null
  ArmorId: number
  MobilityId: number
  MainSensorId: number
  ExtraSensorId: number
  StealthOverride: number
  Ability1Id: number
  Ability2Id: number
  Ability3Id: number
  // Turret0Id .. Turret20Id accessed dynamically
  [key: `Turret${number}Id`]: number
}